Fix the collision base system
Rectangles of the snake part and snake head was indistinguishable because Rectangle is a struct and even that they are different variables since they have the same values they are equal. This causes mistakes in our collision detection. Now Rectangle SnakeHead property is deleted. Instead last item of the list SnakeParts is the head.This is also more intutive