Make the snake grow when eating food
Fixed the bug where the rectangle of the snake was not updating in sync with the Positions of the snake. Actually it would be a better implementation of the concept if snake parts where simply classes which had both rectangle and position fields. That would be easier to read but also heavier with the cpu overload.