Add intro screen. Create a button class. Add a button to intro
Input Manager now checks if LMB is clicked and records the state and triggers an event. Every time that LMB is clicked our button class checks whether the cursor is on the button rectangle. If it is it changes the gamestate to 'PLAY'. Game is hold as a property in static global class to easy access to GameManager. GameManager property in Game1 is also made a public get private set property for the same reason